add all the minimal plan nodes

+++ optd/migrations/2025-01-22-223940_create_logical_exprs/up.sql
@@ -4,7 +4,7 @@ CREATE TABLE logical_exprs (
     -- The logical expression id.
     id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
     -- The logical operator descriptor id.
-    logical_op_desc_id INTEGER NOT NULL,
+    logical_op_desc_id BIGINT NOT NULL,
     -- The group this logical expression belongs to.
     group_id BIGINT NOT NULL, -- groups.id
     -- Time at which the logical expression is created.
